fixed int sizes in powers

+++ HW02 - Intro to C/powers.c
@@ -3,21 +3,23 @@
 
 // not in header file because this is implementation, not use
 // power function
-uint8_t p (uint8_t a, uint8_t b) {
-    uint8_t retval = 1;
-    for (uint8_t i = 0; i < b; i++) {
+
+// REDO: Larger integers
+uint32_t p (uint32_t a, uint32_t b) {
+    uint32_t retval = 1;
+    for (uint32_t i = 0; i < b; i++) {
         retval = retval * a;
     }
     return retval;
 }
 
-uint8_t powers (uint8_t x) {
+uint32_t powers (uint32_t x) {
     // two is minimum value per instructions
-    uint8_t a = 2;
-    uint8_t b = 2;
-    uint8_t number_of_pairs = 0;
+    uint32_t a = 2;
+    uint32_t b = 2;
+    uint32_t number_of_pairs = 0;
     while (1) {
-        uint8_t a_power_b = p(a, b);
+        uint32_t a_power_b = p(a, b);
         // iterative condition
         if (a_power_b <= x) {
             number_of_pairs++;
